# openapi/monitorcontrol.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MonitorControl HTTP API",
    "description": "SDAP discovery and SDCP/VMC/VMS/VMA control. Firmware routes require configuration + X-Firmware-Ack: CONFIRM.",
    "version": "v1"
  },
  "paths": {
    "/api/health": {
      "get": {
        "tags": [
          "monitor"
        ],
        "operationId": "Health",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/sdap/discover": {
      "get": {
        "tags": [
          "monitor"
        ],
        "operationId": "SdapDiscover",
        "parameters": [
          {
            "name": "durationMs",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "bind",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/vmc/get": {
      "post": {
        "tags": [
          "monitor"
        ],
        "operationId": "VmcGet",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VmcGetBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/vmc/set": {
      "post": {
        "tags": [
          "monitor"
        ],
        "operationId": "VmcSet",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VmcSetBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/vmc/broadcast": {
      "post": {
        "tags": [
          "monitor"
        ],
        "operationId": "VmcBroadcast",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VmcBroadcastBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/vms/product-info": {
      "post": {
        "tags": [
          "monitor"
        ],
        "operationId": "VmsProductInfo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HostBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/vma/control-software-version": {
      "post": {
        "tags": [
          "monitor"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HostBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/vma/kernel-version": {
      "post": {
        "tags": [
          "monitor"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HostBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/vma/rtc": {
      "post": {
        "tags": [
          "monitor"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HostBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/vma/fpga1-version": {
      "post": {
        "tags": [
          "monitor"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HostBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/vma/fpga2-version": {
      "post": {
        "tags": [
          "monitor"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HostBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/vma/fpga-core-version": {
      "post": {
        "tags": [
          "monitor"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HostBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/vma/firmware/upgrade-kernel-size": {
      "post": {
        "tags": [
          "monitor"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FirmwareSizeBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/vma/firmware/upgrade-fpga-size": {
      "post": {
        "tags": [
          "monitor"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FirmwareSizeBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/vma/firmware/upgrade-chunk": {
      "post": {
        "tags": [
          "monitor"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FirmwareChunkBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/vma/firmware/upgrade-restart": {
      "post": {
        "tags": [
          "monitor"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HostBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/events/monitor": {
      "get": {
        "tags": [
          "monitor"
        ],
        "description": "Server-Sent Events: JSON snapshots of STATget fields (polling SDCP on the server).",
        "operationId": "MonitorEventsSse",
        "parameters": [
          {
            "name": "host",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fields",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "intervalMs",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "FirmwareChunkBody": {
        "type": "object",
        "properties": {
          "host": {
            "type": "string",
            "nullable": true
          },
          "timeoutMs": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "chunkIndex": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "FirmwareSizeBody": {
        "type": "object",
        "properties": {
          "host": {
            "type": "string",
            "nullable": true
          },
          "timeoutMs": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "sizeBytes": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "HostBody": {
        "type": "object",
        "properties": {
          "host": {
            "type": "string",
            "nullable": true
          },
          "timeoutMs": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "VmcBroadcastBody": {
        "type": "object",
        "properties": {
          "scope": {
            "type": "string",
            "nullable": true
          },
          "groupId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "broadcastAddress": {
            "type": "string",
            "nullable": true
          },
          "port": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "localBind": {
            "type": "string",
            "nullable": true
          },
          "tokens": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "vmcItem": {
            "type": "string",
            "description": "Optional: B000|monitor (default) or B001|builtIn for SDCP item number in the VMC header.",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "VmcGetBody": {
        "type": "object",
        "properties": {
          "host": {
            "type": "string",
            "nullable": true
          },
          "field": {
            "type": "string",
            "nullable": true
          },
          "timeoutMs": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "sdcpUnitId": {
            "type": "integer",
            "format": "int32",
            "description": "Optional 0–255: SDCP single-connection unit (group 0). Omit for P2P (0,0).",
            "nullable": true
          },
          "vmcItem": {
            "type": "string",
            "description": "Optional: B000|monitor (default) or B001|builtIn.",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "VmcSetBody": {
        "type": "object",
        "properties": {
          "host": {
            "type": "string",
            "nullable": true
          },
          "args": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "timeoutMs": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "sdcpUnitId": {
            "type": "integer",
            "format": "int32",
            "description": "Optional 0–255: SDCP single-connection unit (group 0). Omit for P2P (0,0).",
            "nullable": true
          },
          "vmcItem": {
            "type": "string",
            "description": "Optional: B000|monitor (default) or B001|builtIn.",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}